Add support for `.alexrc`, `package.json` config
This feature adds support for configuration through `.alexrc` files,
and a `alex` field in `package.json` files.

To kick things off, an `allow` array can be specified in these
configurations, which holds a list of rules to be ignored when
detecting suggestions.

For example, the following can be used inside a `package.json` file to
not warn about `butt`:

```txt
{
  ...
  "alex": {
    "allow": ["butt"]
  },
  ...
}
```

...or, the following to not warn about `boogeyman` in an `.alexrc`
file:

```json
{
  "allow": ["boogeyman-boogeywoman"]
}
```

Closes GH-18.